rename: move inner class to upper level (#9)
