+++ src/test/java/com/shyoon/wms/location/feature/StorageType.java
@@ -1,0 +1,11 @@
+package com.shyoon.wms.location.feature;
+
+enum StorageType {
+    TOTE("토트바구니");
+
+    private final String description;
+
+    StorageType(String description) {
+        this.description = description;
+    }
+}
