+++ src/test/java/com/shyoon/wms/location/feature/UsagePurpose.java
@@ -1,0 +1,11 @@
+package com.shyoon.wms.location.feature;
+
+enum UsagePurpose {
+    MOVE("이동 목적");
+
+    private final String description;
+
+    UsagePurpose(String description) {
+        this.description = description;
+    }
+}
